+++ src/watney_bringup/params/lifecycle_mgr.yaml
@@ -1,0 +1,11 @@
+# config/lifecycle_mgr.yaml
+#
+# Example configuration file to automatically handle the ldlidar lifecycle node
+
+/**:
+  ros__parameters:
+    autostart: True
+    node_names: ['ldlidar_node']
+    bond_timeout: 5.0
+    attempt_respawn_reconnection: true
+    bond_respawn_max_duration: 10.0
