+++ src/watney_bringup/urdf/ldlidar_descr.urdf.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0"?>
+
+<robot name="ldlidar">
+
+    <link name="ldlidar_base">
+        <visual>
+            <origin xyz="0 0 0" rpy="0 0 0"/>
+            <geometry>
+                <mesh filename="package://ldlidar_node/urdf/LD19.stl" scale="0.001 0.001 0.001" />
+            </geometry>
+            <material name="ldlidar_mat">
+                <color rgba="0.8 0.8 0.8 0.8"/>
+            </material>
+        </visual>
+    </link>
+
+    <joint name="ldlidar_link_joint" type="fixed">
+        <parent link="ldlidar_base"/>
+        <child link="ldlidar_link"/>
+        <origin xyz="0 0 0.02745" rpy="0 0 0" />
+    </joint>
+
+    <link name="ldlidar_link" />
+
+</robot>
